+++ src/FileHelper/FileInfo/FileInfoException.php
@@ -16,5 +16,8 @@ use AppUtils\FileHelper_Exception;
  */
 class FileInfoException extends FileHelper_Exception
 {
-    public const ERROR_INVALID_OVERRIDE_CLASS = 170401;
+    public const int ERROR_INVALID_OVERRIDE_CLASS = 170401;
+    public const int ERROR_CANNOT_RENAME_TARGET_EXISTS = 170402;
+    public const int ERROR_FAILED_TO_RENAME_PATH = 170403;
+    public const int ERROR_CANNOT_RENAME_WITH_PATH = 170404;
 }